+++ pkg/githubclient/client_repository.go
@@ -143,13 +143,16 @@ func (c *Client) SetRepositoryPermissions(ctx context.Context, repository *githu
 		c.logger.Debugf("granted %q permission to %q", permission, teamSlug)
 	}
 
-	input := &github.DefaultWorkflowPermissionRepository{
-		DefaultWorkflowPermissions: github.Ptr("write"),
+	if !c.dryRun {
+		input := &github.DefaultWorkflowPermissionRepository{
+			DefaultWorkflowPermissions: github.Ptr("write"),
+		}
+		_, _, err := underlyingClient.Repositories.UpdateDefaultWorkflowPermissions(ctx, owner, repo, *input)
+		if err != nil {
+			return microerror.Mask(err)
+		}
 	}
-	_, _, err := underlyingClient.Repositories.UpdateDefaultWorkflowPermissions(ctx, owner, repo, *input)
-	if err != nil {
-		return microerror.Mask(err)
-	}
+	c.logger.Debug("set default workflow permissions to write")
 
 	c.logger.Debug("granted permission on repository")